Add build script
Add build script using bash that concatenates the files, modifies
them for one single export then runs babel and minify.

The the files needs to be sorted in dependency order, the infrastrucutre
is in place but no sorting is done yet. Currently the only sorting made
is that files without dependencies comes first.

The dist folder has been added to gitignore.