Make sure all functions are curried
An extra test was added to the testrunner. Non functions and curry
itself are ignored.

The test checks that the result of calling the function with no
arguments is the function itself since this is a property of a curried
function and not expected to be true for any other function.